+++ src/components/layout/index.js
@@ -1,2 +1,3 @@
 export * from "./Header";
 export * from "./Footer";
+export * from "./MarkDownContent";
