+++ src/components/main/index.js
@@ -1,3 +1,2 @@
 export * from "./RecentPostLists";
 export * from "./PostLists";
-export * from "./CategoryFilter";
