+++ src/pages/LoginPage.jsx
@@ -130,105 +130,6 @@ const LoginTitle = styled.h1`
   margin-bottom: 2rem;
 `;
 
-const LoginForm = styled.form`
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-`;
-
-const InputGroup = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-const LoginInput = styled.input`
-  width: 100%;
-  height: 50px;
-  font: var(--Body-R);
-  color: var(--Text-Color);
-  background-color: var(--Back-Color);
-  border: 2px solid var(--Border-Color);
-  border-radius: 8px;
-  padding: 0 1rem;
-  transition: all 0.2s ease;
-
-  &:focus {
-    outline: none;
-    border-color: var(--Brand-Colors);
-    box-shadow: 0 0 0 3px rgb(59 130 246 / 0.1);
-  }
-
-  &::placeholder {
-    color: var(--Text-Color);
-    opacity: 0.6;
-  }
-`;
-
-const ErrorMessage = styled.div`
-  color: #ef4444;
-  font: var(--Caption-R);
-  text-align: center;
-  margin-top: -0.5rem;
-`;
-
-const LoginButton = styled.button`
-  width: 100%;
-  height: 50px;
-  background-color: var(--Brand-Colors);
-  color: white;
-  font: var(--Body-B);
-  border: none;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  margin-top: 0.5rem;
-
-  &:hover:not(:disabled) {
-    background-color: #2563eb;
-    transform: translateY(-1px);
-  }
-
-  &:disabled {
-    opacity: 0.6;
-    cursor: not-allowed;
-    transform: none;
-  }
-
-  &:active {
-    transform: translateY(0);
-  }
-`;
-
-const ToggleSection = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  margin-top: 1.5rem;
-  padding-top: 1.5rem;
-  border-top: 1px solid var(--Border-Color);
-`;
-
-const ToggleText = styled.span`
-  font: var(--Body-R);
-  color: var(--Text-Color);
-  opacity: 0.7;
-`;
-
-const ToggleButton = styled.button`
-  font: var(--Body-B);
-  color: var(--Brand-Colors);
-  background: none;
-  border: none;
-  cursor: pointer;
-  transition: all 0.2s ease;
-
-  &:hover {
-    text-decoration: underline;
-    color: #2563eb;
-  }
-`;
-
 const BackButton = styled.button`
   width: 100%;
   height: 40px;
